vector-kb/requirements.txt
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11